backend/requirements.txt
#cd backend
#pip install -r requirements.txt

fastapi
uvicorn[standard]
pymongo
python-multipart
python-dotenv
numpy
pandas
scikit-learn
serial
motor